ADM-1000 [frontend]: refactor: rename the deployment frequency settings component
