+++ frontend/__tests__/containers/MetricsStep/DeploymentFrequencySettings/BranchSelection.test.tsx
@@ -1,5 +1,5 @@
 import { MOCK_SOURCE_CONTROL_VERIFY_BRANCH_URL, MOCK_SOURCE_CONTROL_VERIFY_REQUEST_PARAMS } from '@test/fixtures';
-import { BranchSelection } from '@src/containers/MetricsStep/DeploymentFrequencySettings/BranchSelection';
+import { BranchSelection } from '@src/containers/MetricsStep/PipelineConfiguration/BranchSelection';
 import { updatePipelineToolVerifyResponse, updateSourceControl } from '@src/context/config/configSlice';
 import { render, screen, waitFor } from '@testing-library/react';
 import { setupStore } from '@test/utils/setupStoreUtil';
