+++ frontend/src/containers/MetricsStep/PipelineConfiguration/BranchSelection/BranchChip/index.tsx
