+++ frontend/src/containers/MetricsStep/PipelineConfiguration/PipelineMetricSelection/index.tsx
@@ -17,9 +17,9 @@ import {
   selectPipelineList,
 } from '@src/context/config/configSlice';
 
-import { FormControlWrapper } from '@src/containers/MetricsStep/DeploymentFrequencySettings/SingleSelection/style';
-import { SingleSelection } from '@src/containers/MetricsStep/DeploymentFrequencySettings/SingleSelection';
-import { BranchSelection } from '@src/containers/MetricsStep/DeploymentFrequencySettings/BranchSelection';
+import { FormControlWrapper } from '@src/containers/MetricsStep/PipelineConfiguration/SingleSelection/style';
+import { SingleSelection } from '@src/containers/MetricsStep/PipelineConfiguration/SingleSelection';
+import { BranchSelection } from '@src/containers/MetricsStep/PipelineConfiguration/BranchSelection';
 import { ButtonWrapper, PipelineMetricSelectionWrapper, RemoveButton, WarningMessage } from './style';
 import { WarningNotification } from '@src/components/Common/WarningNotification';
 import { useGetMetricsStepsEffect } from '@src/hooks/useGetMetricsStepsEffect';
